requirements.txt
# GarageIQ Production Dependencies

# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.4.0
streamlit>=1.25.0

# LLM Integration
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.1.0
langchain-groq>=0.0.1
groq>=0.4.0

# Translation & NLP
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99
langdetect>=1.0.9
sentence-transformers>=2.2.0
chromadb>=0.4.15

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Logging & Monitoring
python-json-logger>=2.0.0
prometheus-client>=0.17.0

# UI & Reporting
audio-recorder-streamlit>=0.0.10
reportlab>=4.0.0
